Update poliza alert and add string formatter file
