+++ src/app/modules/private/documents/remaining-documents/remaining-documents.component.ts
@@ -15,6 +15,7 @@ import { FileValidatorDirective } from 'src/app/directives/file-validator.direct
 import { PipesModule } from 'src/app/pipes/pipes.module';
 import { AlertService } from 'src/app/services/alerts/alert.service';
 import { FormUtilsService } from 'src/app/services/form-utils/form-utils.service';
+import { formatListWithY } from 'src/app/utils/string-utils';
 
 @Component({
    selector: 'app-remaining-documents',
@@ -127,7 +128,7 @@ export class RemainingDocumentsComponent implements OnInit {
     if (onFocus && this.hasShownAmountMessage) return;
     this.hasShownAmountMessage = true;
 
-    // Group by SMLVs num max
+    // Group by SMLVs num min
     const groups: { [smlvNum: number]: string[] } = {};
     for (let type in this.typePolicyProviderConfig) {
       let smlvNum = this.typePolicyProviderConfig[type];
@@ -142,16 +143,16 @@ export class RemainingDocumentsComponent implements OnInit {
       let value = this.SMLV * parseInt(smlvNum);
       html += `
         <p>
-          Para <strong>${provTypes.join(', ')}</strong> el valor no debe superar el tope de
+          Para <strong>${formatListWithY(provTypes)}</strong>, el valor mínimo requerido de la póliza es equivalente a
           <strong>${smlvNum} SMLV</strong>
-          (<strong>${smlvNum} x ${this.SMLV.toLocaleString('es-CO')} = $${value.toLocaleString('es-CO')}</strong>)
+          (<strong>${smlvNum} x $${this.SMLV.toLocaleString('es-CO')} = $${value.toLocaleString('es-CO')}</strong>)
         </p>
       `;
     }
     html += `<p class="mt-3">Por favor verifica este monto.</p>`;
 
     this.alertService.info(
-      'Información sobre topes de póliza',
+      'Información sobre valores de póliza',
       html,
       {
         nzOkText: 'Entendido',
